refactor(nvisy-ocr): hoist src/ocr/* to crate root; move engine + provider enum to src/engine/
The old src/ocr/ wrapper module added a redundant layer — everything
in this crate is OCR-related, so the prefix was pure noise.

Layout
- src/lib.rs declares pub mod backend / engine / provider and
  hoists the common types (Backend, OcrEngine, OcrProvider,
  ImageInput, etc.) to the crate root.
- src/backend/    (was src/ocr/backend/): trait + I/O types.
- src/engine/mod.rs (was src/ocr/mod.rs): OcrEngine type.
- src/engine/provider.rs (was src/ocr/params.rs): OcrProvider enum
  that resolves config into a concrete Backend. Renamed from
  params because the file is the provider enum, not param structs.
- src/provider/   (was src/ocr/provider/): concrete backends.

Internal `crate::ocr::backend::*` / `crate::ocr::provider::*` paths
rewritten to `crate::backend::*` / `crate::provider::*`. Tracing
target renamed from "nvisy_ocr::ocr" to "nvisy_ocr::engine".

No public-API change: consumers used the crate-root re-exports
(`use nvisy_ocr::{OcrEngine, OcrProvider, ImageInput, ...}`)
which are preserved.
