+++ crates/nvisy-ocr/src/backend/output.rs
