+++ crates/nvisy-ocr/src/provider/aws_textract/params.rs
