+++ crates/nvisy-ocr/src/provider/azure_docai/mod.rs
