+++ crates/nvisy-ocr/src/provider/datalab_surya/params.rs
