+++ crates/nvisy-ocr/src/provider/google_vision/params.rs
