+++ crates/nvisy-ocr/src/provider/paddle_paddlex/mod.rs
