+++ crates/nvisy-ocr/src/provider/paddle_paddlex/params.rs
